Target SDK 32 instead of the Tiramisu preview
It turns out apps built with the preview SDK cannot be installed on
older stable versions of Android.

With BCR now targeting SDK 32, the creation of the notification channel
has been moved from RecorderInCallService to RecorderApplication to
ensure that Android 13 Beta devices automatically prompt for
notification permissions on the first app launch.

Once the stable SDK 33 is released, we can re-add the runtime
permissions request for POST_NOTIFICATIONS.

Fixes: #1
